fix: improve destPostalCode extraction accuracy
Fixed bug where the destPostalCode would be extracted as the last line of the toAddress including city and province in addition to postal code. This is done by ensuring the text block used for extracting the destPostalCode is restricted to the one only containing the postal code and nothing else.